feat(TCOMP-2107): refine impl
- refactor static classes and localconf getters
- beam tryout
