+++ component-runtime-impl/src/main/java/org/talend/sdk/component/runtime/input/PartitionMapperImpl.java
@@ -15,8 +15,6 @@
  */
 package org.talend.sdk.component.runtime.input;
 
-import static java.util.Collections.emptySet;
-import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toList;
 
 import java.io.ByteArrayInputStream;
@@ -28,22 +26,20 @@ import java.io.Serializable;
 import java.lang.reflect.Method;
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
-import java.util.function.Supplier;
 
 import org.talend.sdk.component.api.input.Assessor;
 import org.talend.sdk.component.api.input.Emitter;
 import org.talend.sdk.component.api.input.Split;
-import org.talend.sdk.component.api.service.configuration.LocalConfiguration;
 import org.talend.sdk.component.runtime.base.Delegated;
 import org.talend.sdk.component.runtime.base.LifecycleImpl;
 import org.talend.sdk.component.runtime.serialization.ContainerFinder;
 import org.talend.sdk.component.runtime.serialization.EnhancedObjectInputStream;
 
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 public class PartitionMapperImpl extends LifecycleImpl implements Mapper, Delegated {
 
     private static final Object[] NO_ARG = new Object[0];
@@ -59,19 +55,6 @@ public class PartitionMapperImpl extends LifecycleImpl implements Mapper, Delega
     private transient Method inputFactory;
 
     private transient Function<Long, Object[]> splitArgSupplier;
-
-    private final Supplier<LocalConfiguration> defaultLocalConfiguration = () -> new LocalConfiguration() {
-
-        @Override
-        public String get(final String key) {
-            return null;
-        }
-
-        @Override
-        public Set<String> keys() {
-            return emptySet();
-        }
-    };
 
     public PartitionMapperImpl(final String rootName, final String name, final String inputName, final String plugin,
             final boolean stream, final Serializable instance) {
@@ -110,70 +93,13 @@ public class PartitionMapperImpl extends LifecycleImpl implements Mapper, Delega
         // to see it is useful before doing it,
         // java 7/8 made enough progress to probably make it smooth OOTB
         final Serializable input = Serializable.class.cast(doInvoke(inputFactory));
+        log.debug("[PartitionMapperImpl#create] isStream? {}.", isStream());
         if (isStream()) {
-            return new StreamingInputImpl(rootName(), inputName, plugin(), input, loadRetryConfiguration(),
-                    loadStopStrategy());
+            return new StreamingInputImpl(rootName(), inputName, plugin(), input,
+                    Streaming.loadRetryConfiguration(plugin()),
+                    Streaming.loadStopStrategy(plugin()));
         }
         return new InputImpl(rootName(), inputName, plugin(), input);
-    }
-
-    private StreamingInputImpl.RetryConfiguration loadRetryConfiguration() {
-        // note: this configuratoin could be read on the mapper too and distributed
-        final LocalConfiguration configuration = ofNullable(ContainerFinder.Instance.get().find(plugin()))
-                .map(it -> it.findService(LocalConfiguration.class))
-                .orElseGet(defaultLocalConfiguration);
-        final int maxRetries = ofNullable(configuration.get("talend.input.streaming.retry.maxRetries"))
-                .map(Integer::parseInt)
-                .orElse(Integer.MAX_VALUE);
-        return new StreamingInputImpl.RetryConfiguration(maxRetries, getStrategy(configuration));
-    }
-
-    private StreamingInputImpl.RetryStrategy getStrategy(final LocalConfiguration configuration) {
-        switch (ofNullable(configuration.get("talend.input.streaming.retry.strategy")).orElse("constant")) {
-        case "exponential":
-            return new StreamingInputImpl.RetryConfiguration.ExponentialBackoff(
-                    ofNullable(configuration.get("talend.input.streaming.retry.exponential.exponent"))
-                            .map(Double::parseDouble)
-                            .orElse(1.5),
-                    ofNullable(configuration.get("talend.input.streaming.retry.exponential.randomizationFactor"))
-                            .map(Double::parseDouble)
-                            .orElse(.5),
-                    ofNullable(configuration.get("talend.input.streaming.retry.exponential.maxDuration"))
-                            .map(Long::parseLong)
-                            .orElse(TimeUnit.MINUTES.toMillis(5)),
-                    ofNullable(configuration.get("talend.input.streaming.retry.exponential.initialBackOff"))
-                            .map(Long::parseLong)
-                            .orElse(TimeUnit.SECONDS.toMillis(1)),
-                    0);
-        case "constant":
-        default:
-            return new StreamingInputImpl.RetryConfiguration.Constant(
-                    ofNullable(configuration.get("talend.input.streaming.retry.constant.timeout"))
-                            .map(Long::parseLong)
-                            .orElse(500L));
-        }
-    }
-
-    private StreamingInputImpl.StopStrategy loadStopStrategy() {
-        final LocalConfiguration configuration = ofNullable(ContainerFinder.Instance.get().find(plugin()))
-                .map(it -> it.findService(LocalConfiguration.class))
-                .orElseGet(defaultLocalConfiguration);
-        final Long maxReadRecords = ofNullable(configuration.get("talend.input.streaming.maxRecords"))
-                .map(Long::parseLong)
-                .orElse(
-                        ofNullable(configuration.get("streaming.maxRecords"))
-                        .map(Long::parseLong)
-                        .orElse(null)
-                                // TODO orElse system property
-                );
-        final Long maxActiveTime = ofNullable(configuration.get("talend.input.streaming.maxDurationMs"))
-                .map(Long::parseLong)
-                .orElse(
-                         ofNullable(configuration.get("streaming.maxDurationMs"))
-                        .map(Long::parseLong)
-                        .orElse(null)
-                );
-        return new StreamingInputImpl.StopConfiguration(maxReadRecords, maxActiveTime, null);
     }
 
     @Override
